Update tests: handle auth changes, handle reading mode in links
