+++ tests/MockCloudflarePurgeService.php
@@ -13,10 +13,13 @@ class MockCloudflarePurgeService extends CloudflarePurgeService implements TestO
 {
 
     private static $enabled = true;
-    private static $email = "test@example.com";
-    private static $auth_key = "KEY-test-123-abcd";
-    private static $auth_token = "TOKEN-test-123-abcd";
+
+    private static $email = "test@example.com";// legacy
+    private static $auth_key = "KEY-test-123-abcd";// legacy
     private static $base_url = '';
+
+    private static $auth_token = "TOKEN-test-123-abcd";
+
     private static $zone_id = "test-zone";
 
     private static $endpoint_base_uri = "";
